Property 'endurance' added to episodes.
This property defined the lifetime of a recorded episode in the database.
It is defined in feed, station or show sections in the config file and defaults
to two weeks.
Run 'recorder feed cleanup' to remove overdue episodes. The database entry as
well as the media file gets removed, when the episode recording start time is
longer ago than the defined endurance.